requirements.txt
streamlit>=1.35.0
openai>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=1.1.0
scikit-learn>=1.3.0
streamlit-agraph>=0.0.45
xlsxwriter>=3.1.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
sweetviz>=2.3.0
ydata-profiling>=4.6.0
pygwalker>=0.4.0
google-cloud-aiplatform>=1.50.0
google-auth>=2.28.0
